requirements.txt
discord.py
sqlite3
asyncio
gspread